+++ R/tar_map_rep_raw.R
@@ -12,6 +12,7 @@
 #'   See the "Target objects" section for background.
 #' @inheritSection tar_map Target objects
 #' @inheritSection tar_rep Replicate-specific seeds
+#' @inheritParams tar_rep
 #' @param command Language object, R code for a single replicate. Must return
 #'   a data frame.
 #' @param names Language object with a tidyselect expression
@@ -74,6 +75,7 @@ tar_map_rep_raw <- function(
   columns = quote(tidyselect::everything()),
   batches = 1,
   reps = 1,
+  rep_workers = 1,
   combine = TRUE,
   tidy_eval = targets::tar_option_get("tidy_eval"),
   packages = targets::tar_option_get("packages"),
@@ -110,6 +112,8 @@ tar_map_rep_raw <- function(
   targets::tar_assert_dbl(reps)
   targets::tar_assert_scalar(combine)
   targets::tar_assert_lgl(combine)
+  tar_assert_rep_workers(rep_workers)
+  rep_workers <- as.integer(rep_workers)
   envir <- targets::tar_option_get("envir")
   command <- tar_raw_command(name, command)
   command <- targets::tar_tidy_eval(as.expression(command), envir, tidy_eval)
@@ -137,6 +141,7 @@ tar_map_rep_raw <- function(
     command = command,
     name_batch = name_batch,
     reps = reps,
+    rep_workers = rep_workers,
     iteration = "vector"
   )
   target_dynamic <- targets::tar_target_raw(