+++ R/tar_render_rep_raw.R
@@ -42,9 +42,9 @@
 #'   See the "Target objects" section for background.
 #' @inheritSection tar_map Target objects
 #' @inheritSection tar_rep Replicate-specific seeds
-#' @inheritSection tar_rep Nested futures for batched replication
 #' @inheritParams targets::tar_target
 #' @inheritParams rmarkdown::render
+#' @inheritParams tar_rep
 #' @param path Character string, file path to the R Markdown source file.
 #'   Must have length 1.
 #' @param params Expression object with code to generate
@@ -102,6 +102,7 @@ tar_render_rep_raw <- function(
   path,
   params = expression(NULL),
   batches = NULL,
+  rep_workers = 1,
   packages = targets::tar_option_get("packages"),
   library = targets::tar_option_get("library"),
   format = targets::tar_option_get("format"),
@@ -130,6 +131,8 @@ tar_render_rep_raw <- function(
     names(args %||% list(x = 1)),
     "args must be a named list."
   )
+  tar_assert_rep_workers(rep_workers)
+  rep_workers <- as.integer(rep_workers)
   name_params <- paste0(name, "_params")
   sym_params <- as.symbol(name_params)
   target_params <- targets::tar_target_raw(
@@ -150,7 +153,7 @@ tar_render_rep_raw <- function(
   )
   target <- targets::tar_target_raw(
     name = name,
-    command = tar_render_rep_command(name, path, quiet, args),
+    command = tar_render_rep_command(name, path, quiet, args, rep_workers),
     pattern = substitute(map(x), env = list(x = sym_params)),
     packages = packages,
     library = library,
@@ -226,14 +229,21 @@ tar_render_rep_run_params <- function(params, batches) {
   params
 }
 
-tar_render_rep_command <- function(name, path, quiet, args) {
+tar_render_rep_command <- function(name, path, quiet, args, rep_workers) {
   args$input <- path
   args$knit_root_dir <- quote(getwd())
   args$quiet <- quiet
   params <- as.symbol(paste0(name, "_params"))
   deps <- call_list(as_symbols(knitr_deps(path)))
   fun <- call_ns("tarchetypes", "tar_render_rep_run")
-  exprs <- list(fun, path = path, params = params, args = args, deps = deps)
+  exprs <- list(
+    fun,
+    path = path,
+    params = params,
+    args = args,
+    deps = deps,
+    rep_workers = rep_workers
+  )
   as.expression(as.call(exprs))
 }
 
@@ -246,11 +256,12 @@ tar_render_rep_command <- function(name, path, quiet, args) {
 #' @return Character vector with the path to the R Markdown
 #'   source file and the rendered output file. Both paths
 #'   depend on the input source path, and they have no defaults.
+#' @inheritParams tar_rep
 #' @param path Path to the R Markdown source file.
 #' @param args A named list of arguments to `rmarkdown::render()`.
 #' @param deps An unnamed list of target dependencies of the R Markdown
 #'   report, automatically created by `tar_render_rep()`.
-tar_render_rep_run <- function(path, params, args, deps) {
+tar_render_rep_run <- function(path, params, args, deps, rep_workers) {
   targets::tar_assert_package("rmarkdown")
   rm(deps)
   gc()
@@ -258,36 +269,81 @@ tar_render_rep_run <- function(path, params, args, deps) {
   args$envir <- args$envir %|||% targets::tar_envir(default = envir)
   force(args$envir)
   params <- split(params, f = seq_len(nrow(params)))
-  out <- furrr::future_map(
-    .x = seq_along(params),
-    .f = ~tar_render_rep_rep(
-      rep = .x,
-      path = path,
-      params = params,
-      args = args
-    ),
-    .options = furrr::furrr_options(seed = TRUE)
+  call <- quote(
+    function(.x, .y, args, path, seeds) {
+      tarchetypes::tar_render_rep_rep(
+        rep = .x,
+        params = .y,
+        args = args,
+        path = path,
+        seeds = seeds
+      )
+    }
   )
-  unname(unlist(out))
-}
-
-tar_render_rep_rep <- function(rep, path, params, args) {
-  withr::local_options(list(crayon.enabled = NULL))
+  fun <- eval(call, envir = targets::tar_option_get("envir"))
   pedigree <- targets::tar_definition()$pedigree
   name <- pedigree$parent
   batch <- pedigree$index
   reps <- length(params)
-  seed <- produce_seed_rep(name = name, batch = batch, rep = rep, reps = reps)
-  params <- params[[rep]]
+  seeds <- produce_batch_seeds(name = name, batch = batch, reps = reps)
+  if (rep_workers > 1L) {
+    plan_old <- future::plan()
+    on.exit(future::plan(plan_old, .cleanup = FALSE))
+    future::plan(future.callr::callr, workers = rep_workers, .cleanup = FALSE)
+    out <- furrr::future_map2(
+      .x = seq_along(params),
+      .y = params,
+      .f = fun,
+      .options = furrr::furrr_options(
+        seed = 1L,
+        packages = targets::tar_definition()$command$packages,
+        globals = names(targets::tar_option_get("envir"))
+      ),
+      args = args,
+      path = path,
+      seeds = seeds
+    )
+  } else {
+    out <- map2(
+      x = seq_along(params),
+      y = params,
+      f = fun,
+      args = args,
+      path = path,
+      seeds = seeds
+    )
+  }
+  unname(unlist(out))
+}
+
+#' @title Run a rep in a `tar_render_rep()`.
+#' @export
+#' @keywords internal
+#' @description Not a user-side function. Do not invoke directly.
+#' @return Output file paths.
+#' @param rep Rep number.
+#' @param params R Markdown parameters.
+#' @param args Arguments to `rmarkdown::render()`.
+#' @param path R Markdown output file.
+#' @param seeds Random number generator seeds of the batch.
+#' @examples
+#' # See the examples of tar_quarto_rep().
+tar_render_rep_rep <- function(rep, params, args, path, seeds) {
+  withr::local_options(list(crayon.enabled = NULL))
   default_path <- tar_render_rep_default_path(path, params)
   args$output_file <- params[["output_file"]] %|||% default_path
   args$params <- params
   args$params[["output_file"]] <- NULL
   args$params[["tar_group"]] <- NULL
   args$intermediates_dir <- fs::dir_create(tempfile())
-  output <- withr::with_seed(
-    seed = seed,
-    code = do.call(rmarkdown::render, args)
+  seed <- as.integer(if_any(anyNA(seeds), NA_integer_, seeds[rep]))
+  output <- if_any(
+    anyNA(seed),
+    do.call(rmarkdown::render, args),
+    withr::with_seed(
+      seed = seed,
+      code = do.call(rmarkdown::render, args)
+    )
   )
   tar_render_paths(output, path)
 }